requirements.txt
pymongo
mysql-connector-python
redis
